+++ src/intelligence_layer/examples/qa/multiple_chunk_retriever_qa.py
@@ -156,10 +156,15 @@ class MultipleChunkRetrieverQa(
         start_indices: list[int] = []
         combined_text = ""
         for i, chunk in enumerate(chunks):
-            combined_text += source_appendix.format(i=i + 1)
             start_indices.append(len(combined_text))
-            combined_text += chunk + "\n\n"
-        return (TextChunk(combined_text.strip()), start_indices)
+
+            c = source_appendix.format(i=i + 1)
+            c += chunk + "\n\n"
+            c = c.strip()
+            if i != 0:
+                c = " " + c
+            combined_text += c
+        return (TextChunk(combined_text), start_indices)
 
     @staticmethod
     def _get_highlights_per_chunk(
@@ -182,7 +187,10 @@ class MultipleChunkRetrieverQa(
                         end=(
                             highlight.end - current_start
                             if isinstance(next_start, float)
-                            else min(next_start, highlight.end - current_start)
+                            else min(
+                                next_start - current_start,
+                                highlight.end - current_start,
+                            )
                         ),
                         score=highlight.score,
                     )